+++ infra/modules/gcp-bot-runtime/service.tf
@@ -29,7 +29,8 @@ data "google_project" "current" {
 locals {
   bootstrap_image = "us-docker.pkg.dev/cloudrun/container/hello"
 
-  discord_secret_path = "projects/${data.google_project.current.number}/secrets/${google_secret_manager_secret.discord_bot_secrets.secret_id}/versions/latest"
+  discord_secret_path          = "projects/${data.google_project.current.number}/secrets/${google_secret_manager_secret.discord_bot_secrets.secret_id}/versions/latest"
+  valheim_password_secret_path = "projects/${data.google_project.current.number}/secrets/${var.valheim_password_secret_id}/versions/latest"
 }
 
 resource "google_cloud_run_v2_service" "bot" {
@@ -122,6 +123,16 @@ resource "google_cloud_run_v2_service" "bot" {
         name  = "VALHEIM_ZONE"
         value = local.vm_zone
       }
+
+      env {
+        name  = "VALHEIM_STATUS_HTTP_PORT"
+        value = tostring(var.valheim_status_http_port)
+      }
+
+      env {
+        name  = "VALHEIM_PASSWORD_SECRET_PATH"
+        value = local.valheim_password_secret_path
+      }
     }
   }
 
@@ -151,6 +162,7 @@ resource "google_cloud_run_v2_service" "bot" {
 
   depends_on = [
     google_secret_manager_secret_iam_member.bot_can_read_discord_secrets,
+    google_secret_manager_secret_iam_member.bot_can_read_valheim_password,
     google_compute_instance_iam_member.bot_can_admin_valheim_vm,
   ]
 }